Move gnupg installation to Dockerfile
We use gnupg for apt key verification so we should install it once to
allow further streamlining of the build process.

We currently install it and remove it multiple times, and because it is
not a part of the base build, we have to do additional apt processes
just to install it before we add custom repos.